Divided register and login into two separate components
