+++ frontend/src/components/Login/Login.tsx
@@ -1,0 +1,65 @@
+import { Container, Header, Input, InputContainer, InputDescription, Inputs, Question, RedirectContainer, RedirectLink, StyledForm, Submit, SubmitContainer, Text, Underline, Wrapper } from "../../shared/Signup-Login.styled";
+import React from "react";
+import { useState } from "react";
+import { LoggedUser } from "../../types/User";
+import axios from "axios";
+
+export const Login = () => {
+  const [formData, setFormData] = useState<LoggedUser>({
+    username: '',
+    password: '',
+  }); //podaci iz forma
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target;
+    setFormData({ ...formData, [name]: value });
+  };
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    try {
+      const url = 'http://localhost:5185/api/account/login';
+      console.log(formData)
+      const response = await axios.post(url, formData);
+      console.log(response.data);
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  return (
+    <>
+      <Wrapper>
+        <Container>
+          <StyledForm onSubmit={handleSubmit}>
+            <Header>
+                <Text>Login</Text>
+                <Underline></Underline>
+            </Header>
+            <Inputs>
+              <InputContainer>
+                <InputDescription>Username</InputDescription>
+                <Input>
+                  <input type="text" name="username" placeholder="John Doe" onChange={handleChange}/>
+                </Input>
+              </InputContainer>
+              <InputContainer>
+                <InputDescription>Password</InputDescription>
+                <Input>
+                  <input type="password" name="password" placeholder="testpw123" onChange={handleChange}/>
+                </Input>
+              </InputContainer>
+            </Inputs>
+            <RedirectContainer>
+            <Question>Don't have an account?</Question>
+            <RedirectLink href="/register">Register</RedirectLink>
+            </RedirectContainer>
+            <SubmitContainer>
+              <Submit type="submit">Login</Submit>
+            </SubmitContainer>
+          </StyledForm>
+        </Container>
+      </Wrapper>
+    </>
+  );
+}
