+++ frontend/src/pages/RegisterPage.tsx
@@ -1,0 +1,13 @@
+import { Navbar } from '../components/Navbar/Navbar';
+import { Register } from '../components/Register/Register'
+
+function Register_Login() {
+    return (
+        <>
+            <Navbar />
+            <Register />
+        </>
+    );
+}
+
+export default Register_Login;
